Working app v.1

+++ src/main/java/com/lemsviat/lemhomework252/dao/FilesRepository.java
@@ -3,5 +3,5 @@ package com.lemsviat.lemhomework252.dao;
 import com.lemsviat.lemhomework252.entity.File;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FilesRepository extends JpaRepository<File,Integer>{
+public interface FilesRepository extends JpaRepository<File,Long>{
 }
